Restructure network graph
