+++ src/core/node/nodeGraph.ts
@@ -1,3 +1,6 @@
+import * as d3 from 'd3';
+
+import { NetworkGraph } from '../network/networkGraph';
 import { Node } from './node';
 
 function anglePoint(deg: number, radius: number, y0: number = 0): number[] {
@@ -84,7 +87,7 @@ function getTrianglePoints(radius: number): string {
   return points;
 }
 
-export function getPoints(node: Node, radius: number): string {
+function getPoints(node: Node, radius: number): string {
   switch (node.model.elementType) {
     case 'stimulator':
       return getHexagonPoints(radius);
@@ -98,5 +101,294 @@ export function getPoints(node: Node, radius: number): string {
       }
     default:
       return getSquarePoints(radius);
+  }
+}
+
+export class NodeGraph {
+  private _connectorRadius: number = 6;
+  private _networkGraph: NetworkGraph;
+  private _nodeRadius: number = 20;
+  private _selector: d3.Selection<any, any, any, any>;
+
+  constructor(networkGraph: NetworkGraph) {
+    this._networkGraph = networkGraph;
+    this._selector = networkGraph.selector.select('g#nodes');
+  }
+
+  get nodeRadius(): number {
+    return this._nodeRadius;
+  }
+
+  /**
+   * Initialize node graph.
+   *
+   * @remark It is only been executed in networkGraph.init().
+   * For changing in node graph, use update().
+   */
+  init(
+    node: Node,
+    idx: number,
+    elements: SVGGElement[] | ArrayLike<SVGGElement>
+  ): void {
+    // console.log('Init node graph');
+    const elem: d3.Selection<any, any, any, any> = d3.select(elements[idx]);
+    elem.selectAll('*').remove();
+
+    const connector: d3.Selection<any, any, any, any> = elem
+      .append('g')
+      .attr('class', 'connector')
+      .on('mousedown.drag', null);
+
+    connector
+      .append('path')
+      .attr('fill', 'none')
+      .attr('stroke', 'black')
+      .style('opacity', 0)
+      .attr('stroke-width', 16);
+
+    connector
+      .append('path')
+      .attr('class', 'color')
+      .attr('fill', 'none')
+      .attr('stroke-width', this._networkGraph.strokeWidth)
+      .style('pointer-events', 'none');
+
+    const connectorEnd = connector.append('g').attr('class', 'end');
+
+    connectorEnd
+      .append('circle')
+      .attr('class', 'color')
+      .attr('fill', 'white')
+      .attr('r', '6px')
+      .attr('stroke-width', this._networkGraph.strokeWidth)
+      .style('cursor', 'pointer')
+      .on('mousedown', (e: MouseEvent) => {
+        this._networkGraph.reset();
+        this._networkGraph.network.view.selectedNode = node;
+        this._networkGraph.state.enableConnection = true;
+        this._networkGraph.dragLine(e);
+        this._networkGraph.update();
+      });
+
+    // Connector plus symbol made of lines (white lines for spacing):
+    // hline white
+    // coordinates with current config: x1: 30, y1: 25.5, x2: 39.5, y2: 25.5
+    connectorEnd
+      .append('line')
+      .attr('stroke-width', 4)
+      .attr('stroke', 'white')
+      .attr('x1', this._connectorRadius / 3)
+      .attr('x2', (23 / 12) * this._connectorRadius)
+      .attr('y1', -(13 / 12) * this._connectorRadius)
+      .attr('y2', -(13 / 12) * this._connectorRadius)
+      .style('pointer-events', 'none');
+    // vline white
+    // coordinates with current config: x1: 35, y1: 21.5, x2: 35, y2: 31
+    connectorEnd
+      .append('line')
+      .attr('stroke-width', 4)
+      .attr('stroke', 'white')
+      .attr('x1', (7 / 6) * this._connectorRadius)
+      .attr('x2', (7 / 6) * this._connectorRadius)
+      .attr('y1', -(21 / 12) * this._connectorRadius)
+      .attr('y2', -(1 / 6) * this._connectorRadius)
+      .style('pointer-events', 'none');
+    // hline colored
+    // coordinates with current config: x1: 31.5, y1: 25.5, x2: 38.5, y2: 25.5
+    connectorEnd
+      .append('line')
+      .attr('class', 'color')
+      .attr('stroke-width', 1.25)
+      .attr('x1', (7 / 12) * this._connectorRadius)
+      .attr('x2', (21 / 12) * this._connectorRadius)
+      .attr('y1', -(13 / 12) * this._connectorRadius)
+      .attr('y2', -(13 / 12) * this._connectorRadius)
+      .style('pointer-events', 'none');
+    // vline colored
+    // coordinates with current config: x1: 35, y1: 22, x2: 35, y2: 29.5
+    connectorEnd
+      .append('line')
+      .attr('class', 'color')
+      .attr('stroke-width', 1.25)
+      .attr('x1', (7 / 6) * this._connectorRadius)
+      .attr('x2', (7 / 6) * this._connectorRadius)
+      .attr('y1', -(5 / 3) * this._connectorRadius)
+      .attr('y2', -(5 / 12) * this._connectorRadius)
+      .style('pointer-events', 'none');
+
+    const soma: d3.Selection<any, any, any, any> = elem
+      .append('g')
+      .attr('class', 'soma');
+
+    if (
+      node.model.elementType == 'neuron' &&
+      node.view.weight == 'inhibitory'
+    ) {
+      soma.append('circle').attr('class', 'shape').attr('r', this._nodeRadius);
+    } else {
+      soma
+        .append('polygon')
+        .attr('class', 'shape')
+        .attr('points', getPoints(node, this._nodeRadius));
+    }
+
+    soma
+      .selectAll('.shape')
+      .style('stroke', node.view.color)
+      .style('stroke-width', this._networkGraph.strokeWidth);
+
+    soma
+      .append('text')
+      .attr('dy', () =>
+        node.model.elementType == 'neuron' && node.view.weight == 'inhibitory'
+          ? '0.4em'
+          : '0.7em'
+      )
+      .text(() => node.view.label);
+
+    soma.on('click', () => {
+      if (
+        this._networkGraph.network.view.selectedNode &&
+        this._networkGraph.state.enableConnection
+      ) {
+        this._networkGraph.cursorPosition = JSON.parse(
+          JSON.stringify(node.view.position)
+        );
+        this._networkGraph.state.connected = true;
+        this._networkGraph.network.connectNodes(
+          this._networkGraph.network.view.selectedNode,
+          node
+        );
+        this._networkGraph.reset();
+        node.view.focus();
+      } else if (this._networkGraph.network.view.selectedNode === node) {
+        this._networkGraph.reset();
+        node.view.focus();
+      } else {
+        this._networkGraph.reset();
+        node.view.select();
+        node.view.focus();
+      }
+      this._networkGraph.state.enableConnection =
+        [18, 225].includes(this._networkGraph.state.keyCode) &&
+        this._networkGraph.network.view.selectedNode;
+      this._networkGraph.update();
+    });
+
+    elem.on('mouseover', () => {
+      this._networkGraph.network.view.focusedNode = node;
+      if (
+        this._networkGraph.network.view.selectedNode &&
+        this._networkGraph.state.enableConnection
+      ) {
+        this._networkGraph.drawLineNodes(
+          this._networkGraph.network.view.selectedNode,
+          node
+        );
+      }
+      this._networkGraph.update();
+    });
+
+    elem.on('mouseout', () => {
+      this._networkGraph.network.view.resetFocus();
+      this._networkGraph.update();
+    });
+  }
+
+  /**
+   * Update node graph.
+   */
+  update() {
+    // console.log('Update node graph');
+    const duration: number = this._networkGraph.state.dragging ? 0 : 250;
+    const t: d3.Transition<any, any, any, any> = d3
+      .transition()
+      .duration(duration);
+
+    // update node group
+    this._selector
+      .selectAll('g.node')
+      .transition(t)
+      .style('opacity', 1)
+      .attr(
+        'transform',
+        (node: Node) =>
+          `translate(${node.view.position.x},${node.view.position.y}) scale( ${
+            node.view.isFocused() ? 1.2 : 1
+          })`
+      );
+
+    // update node shape
+    this._selector
+      .selectAll('g.node')
+      .selectAll('.shape')
+      .style('stroke', (node: Node) => node.view.color)
+      .style(
+        'stroke-width',
+        (node: Node) => (node.n > 1 ? 1.5 : 1) * this._networkGraph.strokeWidth
+      )
+      .style('stroke-dasharray', (node: Node) =>
+        node.view.isSelected() ? '7.85' : ''
+      );
+
+    const connector: d3.Selection<any, any, any, any> =
+      this._selector.selectAll('g.connector');
+
+    connector
+      .transition(t)
+      .style('opacity', (node: Node) =>
+        (node.view.isFocused() || node.view.isSelected()) &&
+        !this._networkGraph.state.enableConnection &&
+        !this._networkGraph.state.dragging
+          ? '1'
+          : '0'
+      );
+
+    // connector animation
+    const connectorEndPos: any = {
+      x: this._nodeRadius + 8,
+      y: this._nodeRadius + 12,
+    };
+
+    connector
+      .selectAll('path')
+      .transition(t)
+      .attr('d', (node: Node) =>
+        this._networkGraph.connectionGraph.drawPath(
+          { x: 0, y: 0 },
+          (node.view.isFocused() || node.view.isSelected()) &&
+            !this._networkGraph.state.enableConnection &&
+            !this._networkGraph.state.dragging
+            ? connectorEndPos
+            : { x: 0, y: 0 },
+          { isTargetMouse: true }
+        )
+      );
+
+    connector
+      .select('.end')
+      .transition(t)
+      .attr('transform', (node: Node) =>
+        (node.view.isFocused() || node.view.isSelected()) &&
+        !this._networkGraph.state.enableConnection &&
+        !this._networkGraph.state.dragging
+          ? `translate(${connectorEndPos.x}, ${connectorEndPos.y})`
+          : 'translate(0,0)'
+      );
+
+    connector
+      .selectAll('line')
+      .transition(t)
+      .attr(
+        'opacity',
+        (node: Node) =>
+          node.view.isFocused() || node.view.isSelected() ? 1 : 0
+        // remove node.view.isSelected() for only hover appearance of the plus symbol
+      );
+
+    connector
+      .selectAll('.color')
+      .transition(t)
+      .style('stroke', (node: Node) => node.view.color);
   }
 }
